requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
Django==3.0
django-crispy-forms==1.8.1
django-generic-json-views==0.8
djangorestframework==3.11.0
gunicorn==20.0.4
idna==2.8
pytz==2019.3
requests==2.22.0
six==1.10.0
sqlparse==0.3.0
urllib3==1.25.7
